requirements.txt
Django==4.0.5
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
django-ninja
